+++ src/gui/style.rs
@@ -202,7 +202,9 @@ pub enum Container {
     Primary,
     ModalForeground,
     ModalBackground,
-    Player,
+    Player {
+        selected: bool,
+    },
     PlayerGroup,
     PlayerGroupControls,
     PlayerGroupTitle,
@@ -220,7 +222,7 @@ impl container::Catalog for Theme {
         container::Style {
             background: Some(match class {
                 Container::Wrapper => Color::TRANSPARENT.into(),
-                Container::Player => self.field.alpha(0.15).into(),
+                Container::Player { .. } => self.field.alpha(0.15).into(),
                 Container::PlayerGroup => self.field.alpha(0.3).into(),
                 Container::PlayerGroupControls => self.field.into(),
                 Container::PlayerGroupTitle => self.field.alpha(0.45).into(),
@@ -232,14 +234,20 @@ impl container::Catalog for Theme {
             border: Border {
                 color: match class {
                     Container::Wrapper => Color::TRANSPARENT,
-                    Container::Player => self.field.alpha(0.8),
+                    Container::Player { selected } => {
+                        if *selected {
+                            self.positive.alpha(0.8)
+                        } else {
+                            self.field.alpha(0.8)
+                        }
+                    }
                     Container::PlayerGroup | Container::PlayerGroupTitle => self.field,
                     Container::PlayerGroupControls => self.disabled,
                     Container::ModalForeground => self.disabled,
                     _ => self.text,
                 },
                 width: match class {
-                    Container::Player
+                    Container::Player { .. }
                     | Container::PlayerGroup
                     | Container::PlayerGroupControls
                     | Container::PlayerGroupTitle
@@ -247,7 +255,9 @@ impl container::Catalog for Theme {
                     _ => 0.0,
                 },
                 radius: match class {
-                    Container::ModalForeground | Container::Player | Container::PlayerGroupControls => 10.0.into(),
+                    Container::ModalForeground | Container::Player { .. } | Container::PlayerGroupControls => {
+                        10.0.into()
+                    }
                     Container::PlayerGroup => Radius::new(10.0).top(0.0),
                     Container::PlayerGroupTitle => Radius::new(10.0).bottom(0.0),
                     Container::ModalBackground => 5.0.into(),